+++ R/helpers.R
@@ -22,3 +22,12 @@ different_lengths <- function(...) {
 duplicate_products <- function(x) {
   any(vapply(x, anyDuplicated, numeric(1L), incomparables = NA) > 0)
 }
+
+#' Check price and quantity arguments
+#' @noRd
+check_pqs <- function(...) {
+  vars <- as.character(match.call()[-1L])
+  if (different_lengths(...)) {
+    stop(gettextf("%s must be the same length", toString(sQuote(vars, "'"))))
+  }
+}