+++ NEWS.md
@@ -4,6 +4,8 @@
 
 - Fixed a bug with `transmute_weights()` where the weights could be negative.
 
+- Price-index functions have better argument checking.
+
 ## Version 0.6.1
 
 - Updated maintainer email.